Made a couple of important changes
In this update I improved the minimap border colors, so that it's more
noticeable on very light themes, but isn't bad on dark themes. I applied
the following value to the viewport_color setting: [102, 102, 102, 50]
Which will effectively make it a gray, transparent, very good for any
color.

* Fixed the minimap border for lighter themes, also looks good in dark
themes now
+~ Made the tabs much smaller, in order to not waste space when you have
tabs with very short names